requirements.txt
flask==2.3.3
requests==2.31.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-data-tables==12.4.3
python-dateutil==2.8.2
waitress==2.1.2
